Remove unnecessary utilities.
In production, these utils literally just return an empty string. In dev, they return the hostname and port of the current page... which is equivalent to an empty string. They also used a commonJS `exports` statement outside of a commonJS environment, which doesn't work.

These utilities were probably more useful when the front and and back end were hosted on different ports.